Full CRUD functionality for `Venues` module.

+++ src/venues/providers/venues.service.ts
@@ -3,9 +3,10 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm/repository/Repository';
 import { Venue } from '../venue.entity';
 import { CreateVenueDto } from '../dtos/create-venue.dto';
-import { GetVenuesDto } from '../dtos/get-venues.dto';
 import { NotFoundException } from '@nestjs/common';
 import { PatchVenueDto } from '../dtos/patch-venue.dto';
+import { GetVenueDetailDto } from '../dtos/get-venue-detail.dto';
+import { plainToInstance } from 'class-transformer';
 
 @Injectable()
 export class VenuesService {
@@ -29,27 +30,41 @@ export class VenuesService {
   }
 
   /* Get a venue by ID */
-  public async getVenueById(id: number): Promise<Venue> {
+  public async getVenueById(id: number): Promise<GetVenueDetailDto> {
     const venue = await this.venueRepository.findOneBy({ id });
     if (!venue) {
       throw new NotFoundException(`Venue with id ${id} not found`);
     }
-    return venue;
+    return plainToInstance(GetVenueDetailDto, venue);
   }
 
   /* Update a venue */
   public async updateVenue(
     id: number,
     updateVenueDto: PatchVenueDto,
-  ): Promise<GetVenuesDto> {
-    const venue = await this.getVenueById(id);
-    Object.assign(venue, updateVenueDto);
-    return this.venueRepository.save(venue);
+  ): Promise<GetVenueDetailDto> {
+    const venue = await this.venueRepository.findOneBy({ id });
+
+    if (!venue) {
+      throw new NotFoundException(`Venue with id ${id} not found`);
+    }
+
+    try {
+      Object.assign(venue, updateVenueDto);
+      const updatedVenue = await this.venueRepository.save(venue);
+      return plainToInstance(GetVenueDetailDto, updatedVenue);
+    } catch (error) {
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+      throw new Error(`Error updating venue: ${error.message}`);
+    }
   }
 
   /* Delete a venue */
   public async deleteVenue(id: number): Promise<void> {
-    const venue = await this.getVenueById(id);
+    const venue = await this.venueRepository.findOneBy({ id });
+    if (!venue) {
+      throw new NotFoundException(`Venue with id ${id} not found`);
+    }
     await this.venueRepository.remove(venue);
   }
 }
